Move SaveFrameAsBMP to shared FrameUtils
