+++ src/shared/FrameLogger.cpp
@@ -5,6 +5,7 @@
 #include <iomanip>
 #include <iostream>
 #include <sstream>
+#include "FrameUtils.h"
 #ifdef _WIN32
 #define NOMINMAX
 #include <windows.h>
@@ -50,60 +51,6 @@ bool FrameLogger::LogFrame(
             << "x" << height << ", " << dataSize << " bytes)" << std::endl;
 
   return true;
-}
-
-bool FrameLogger::SaveFrameAsBMP(
-    uint32_t width,
-    uint32_t height,
-    const std::vector<uint8_t>& frameData,
-    const std::string& filename) {
-#ifdef _WIN32
-  // Create BMP file headers (assumes BGRA format)
-  BITMAPFILEHEADER fileHeader = {};
-  BITMAPINFOHEADER infoHeader = {};
-
-  fileHeader.bfType = 0x4D42; // "BM"
-  fileHeader.bfSize =
-      sizeof(BITMAPFILEHEADER) + sizeof(BITMAPINFOHEADER) + frameData.size();
-  fileHeader.bfOffBits = sizeof(BITMAPFILEHEADER) + sizeof(BITMAPINFOHEADER);
-
-  infoHeader.biSize = sizeof(BITMAPINFOHEADER);
-  infoHeader.biWidth = width;
-  infoHeader.biHeight = -(int)height; // Top-down DIB
-  infoHeader.biPlanes = 1;
-  infoHeader.biBitCount = 32;
-  infoHeader.biCompression = BI_RGB;
-  infoHeader.biSizeImage = frameData.size();
-
-  HANDLE hFile = CreateFileA(
-      filename.c_str(),
-      GENERIC_WRITE,
-      0,
-      nullptr,
-      CREATE_ALWAYS,
-      FILE_ATTRIBUTE_NORMAL,
-      nullptr);
-  if (hFile != INVALID_HANDLE_VALUE) {
-    DWORD written;
-    bool success =
-        WriteFile(hFile, &fileHeader, sizeof(fileHeader), &written, nullptr) &&
-        WriteFile(hFile, &infoHeader, sizeof(infoHeader), &written, nullptr) &&
-        WriteFile(hFile, frameData.data(), frameData.size(), &written, nullptr);
-    CloseHandle(hFile);
-    return success;
-  }
-  return false;
-#else
-  // For non-Windows platforms, save as raw data for now
-  std::ofstream file(filename, std::ios::binary);
-  if (file.is_open()) {
-    file.write(
-        reinterpret_cast<const char*>(frameData.data()), frameData.size());
-    file.close();
-    return true;
-  }
-  return false;
-#endif
 }
 
 void FrameLogger::SaveFramesToDisk() {
