+++ src/shared/FrameUtils.h
@@ -1,6 +1,8 @@
 #pragma once
 #include <chrono>
+#include <cstdint>
 #include <functional>
+#include <string>
 #include <thread>
 #include <vector>
 #include "protocol.h"
@@ -89,3 +91,33 @@ inline bool ReadFrameGeneric(
     return false;
   return true;
 }
+
+#pragma pack(push, 1)
+struct BMPFileHeader {
+  uint16_t bfType{0};
+  uint32_t bfSize{0};
+  uint16_t bfReserved1{0};
+  uint16_t bfReserved2{0};
+  uint32_t bfOffBits{0};
+};
+
+struct BMPInfoHeader {
+  uint32_t biSize{0};
+  int32_t biWidth{0};
+  int32_t biHeight{0};
+  uint16_t biPlanes{1};
+  uint16_t biBitCount{0};
+  uint32_t biCompression{0};
+  uint32_t biSizeImage{0};
+  int32_t biXPelsPerMeter{0};
+  int32_t biYPelsPerMeter{0};
+  uint32_t biClrUsed{0};
+  uint32_t biClrImportant{0};
+};
+#pragma pack(pop)
+
+bool SaveFrameAsBMP(
+    uint32_t width,
+    uint32_t height,
+    const std::vector<uint8_t>& frameData,
+    const std::string& filename);
